crawling/pyproject.toml
[tool.poetry]
name = "crawling"
version = "0.1.0"
description = ""
authors = ["GangBean <xbox2537@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.6.2"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
pandas = "^2.2.1"
openai = "^1.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
